formatters: Use printf for nosyncsh and cross_meson
Portably using echo when escape sequences are involved is more
involved than one would expect.

In particular, the shell that is used to run the generation
step as part of 'docker build' is not guaranteed to be the same
that is used to invoke the install_buildenv function.

For example, bash's echo builtin doesn't handle escape sequences
by default but can be told to do so by passing the -e option;
dash's version of echo, on the other hand, defaults to handling
escape sequences, but when passed the -e option will simply
consider it part of the input and print it back along with the
rest.

bash's behavior results in broken files, with all contents in a
single line, getting generated as part of $CROSS-$NAME-local-env
jobs.

To solve the issue, use printf instead instead of echo. This
command works consistently across shells, and does the right
thing without needing to pass additional flags.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>