facts: mappings: Assume that 'rpm' based distros install 'rpm2cpio'
The 'rpm2cpio' package is part of the 'rpm' package which can be assumed
to be present on RPM based distros.

This fixes the build of the OpenSUSE Leap 15.4 container as they
switched to use rpm-ndb (using 'ndb' as their database backend) instead
of "original" rpm.

Either way the required rpm2cpio program is present.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>