tests: conftest: Add a mechanism to mock whole modules
Sometimes we need a module during runtime, but we don't need that
module for testing and getting it would require installing a bunch of
other useless dependencies in the test environment. For such modules,
we'll create their mock counterparts in test_utils. Let's start with
libvirt and add more in the future.

Signed-off-by: Erik Skultety <eskultet@redhat.com>