config: Drop validation of mandatory keys
With mandatory keys now having some defaults, we don't need the
validation piece of code for them. Effectively this will allow us to
not mandate users to provide a config at all in future patches as long
as they're happy with the defaults (gitlab runner secret being the
exception which is always going to be checked appropriately).

Signed-off-by: Erik Skultety <eskultet@redhat.com>