etc: config: Provide a default root password
This is one of the mandatory keys we have in the config which in turn
requires users to always provide a config.yml in their home even though
they might want to rely on defaults mostly. While setting a secure root
password is for sure a good security practice, most of the VMs created
by lcitool this way are throw away anyway so at least add an option to
use a simple default password which users can change at any time the
same way as it used to be until now.
At the same time though, setting an explicit None password by the user
results in an exception as it should.

This patch aims to speed up the adoption of lcitool for local VMs by
just a tiny bit.

Signed-off-by: Erik Skultety <eskultet@redhat.com>