etc: config: Provide a default value for gitlab.runner_secret
Same as with root_password, this is the last occurrence of a mandatory
key. We used to fail during the config section validation if this key
wasn't provided meaning that because of this we never allowed *not*
having a config in a home directory when dealing with VMs at all.
After this patch there's no more mandatory sections or keys and so
we'll be able to allow users completely rely on our defaults.

As for why we need any default value (like the weird 'NONE' chosen one)
it's because future patches will check the whole merged config against
empty values recursively, i.e. not per-section as it's being done
today. Having YAML set the options as None would instantly fail the
validation which wouldn't help us at all, so a placeholder value is
needed even though it'll never be technically used.

Signed-off-by: Erik Skultety <eskultet@redhat.com>