config: Stop validating the config per-section, do it as a whole
With the mandatory keys gone, we can perform validation of the config
as a whole, we're checking emptyness of values and types anyway - all
per-section value checks have always had to be done afterwards.

This patch re-uses the previous _validate_section helper - now called
_validate_keys - which calls itself recursively to check the whole
config for consistency and type issues.

Signed-off-by: Erik Skultety <eskultet@redhat.com>