ansible: playbooks: bootstrap: Only run python3 install when needed
If run on a distro that doesn't have repos pre-installed, this task
will fail immediately. Some distros will have python3 pre-installed
even though the repos are not there, but a mere check whether python3
is installed will trigger an error in the distro stating that no repos
are available. Let's fail only when we're actually trying to install a
package, so add a check that checks whether python3 is there but
doesn't invoke the package manager. Worth noting is that we need to
ignore all Ansible errors with the python3 checking task because
otherwise Ansible will terminate the playbook if python3 is not
installed on the target node which is what we're intending to fix in
subsequent tasks.

Signed-off-by: Erik Skultety <eskultet@redhat.com>