facts: projects: vm: Drop the 'screen' package from the list
For a generic default setup, this package is totally unnecessary to
have, it's merely a nice-to-have. The bigger problem is that it may not
be available on some platforms/repos which provide a different
alternative. The mere existence of it in the default list can mean
custom/proprietary platforms cannot be prepared by the update playbook
successfully.

Signed-off-by: Erik Skultety <eskultet@redhat.com>