util: DataDir: Introduce a 'path' property
One of the things any consumer of a DataDir instance should be able to
do is to actually get the file system path to the user data directory.
One of such scenarios is allowing users to run their own Ansible tasks
where we need to know where they live. This patch adds a property to
expose the canonical directory path.

Signed-off-by: Erik Skultety <eskultet@redhat.com>