libvirt_wrapper: hosts: Use VIR_DOMAIN_AFFECT_CONFIG to get XML metadata
A bit controversial, but consider having a VM installed already outside
of lcitool. You start it and wish to update it, but you can't because
we only poll the metadata from the current state of the machine's XML.
It is very unlikely someone would inject the correct metadata to just
the runtime XML to get the update going instead of editing the defined
XML with 'virsh edit' which is much more convenient. The thing is, if
the machine is already running, one has to shut it down first so that
the XML changes take effect and only then the update can happen.
In this particular scenario it would make it more convenient to only
query the XML metadata from defined XMLs to build the inventory of
managed hosts, not from the current state. That way, users only need to
update the defined XML, keep the machine running and then immediately
proceed with the update.

Signed-off-by: Erik Skultety <eskultet@redhat.com>