ci: gitlab: Add environment prep scripts for the all mappings test
This patch adds a new ci/gitlab directory mainly to introduce a bunch
of platform specific Bash environment installation bits which we need
before we can run the all_mappings installation test. Many of these are
just symlinks as there are not that big differences in between
individual major releases of a given distros, some couldn't be
symlinked though.

Signed-off-by: Erik Skultety <eskultet@redhat.com>