ci: gitlab: Add a dumb DNF dry-run install wrapper
So DNF actually ends with and error code 1 when doing a dry-run install
which translates to usage of the '--assumeno' option. This means that
for us it's actually indistinguishable between a package name not being
recognized or simply being the result of the operation having been
aborted. We can't use the 'list' command sadly because list doesn't map
tools/keywords/old names to the correct package, e.g. 'dnf list
ebtables' gives an error while 'dnf install ebtables' maps it correctly
to 'iptables-ebtables' on AlmaLinux 8 and CentOS Stream 8.
What this all means is that in order to get all the features of the
'install' command and yet be able to do a dry-run, we need to wrap its
use and provide some dumb parsing over the output and hope for the best
that the output doesn't change in the future.

Signed-off-by: Erik Skultety <eskultet@redhat.com>