.gitlab-ci: Properly test installation of all mappings
This patch reworks the all package installation job so that it actually
tries to install every single mapping we know of on most platforms as
opposed to just merge all projects we currently "maintain" and install
those packages in a single transaction. The obvious problems are:
    - some mappings are never exercised
    - some mappings can actually conflict on a given platform

This patch fixes the above by installing every single of our mappings
(including Python and Perl ones) individually in a dry-run manner so
that we don't waste resources on downloading the packages and writing
to the disk.
There are a few pitfalls though required and ugly solution:
    - for DNF we need to use a wrapper script because DNF doesn't
      implement a proper dry-run install
    - Alpine implements wget in Busybox which cpanm relies on and on
      3.15 it's missing some CLI option so we need to use '--no-wget'
      with cpanm; on 3.16 wget still doesn't implement the option, but
      this one actually works with the defaults and doesn't work at all
      when combined with the '--no-wget' option (it's a mess!)
    - OpenSUSE doesn't create the python3 symlink automatically, it
      only does that with 3.6, so 'update-alternatives' had to be used
      to actually use a more modern Python (lcitool needs >=3.8)
    - all Python installs are happening in a virtual env because Debian
      decided to block pip from installing to system locations

Signed-off-by: Erik Skultety <eskultet@redhat.com>