install: image: Force 0644 permissions on the downloaded vendor images
Libvirt can't reset the ownership on backing images and because these
vendor images are downloaded using Python's temporary files which
hardcode the file permissions to 600, we'd locked ourselves in to
always use sudo to do basically anything to the disk images outside of
libvirt, e.g. using guestfs tools which need to mount the images into
a temporary internal appliance.

Signed-off-by: Erik Skultety <eskultet@redhat.com>