libvirt_wrapper: Add support for specifying a backing volume for create
This whole storage volume creation process we decided to go through
only makes sense when paired with a backing store volume. This can
either be a vendor provided cloud image or a custom template image.
What's important though is that in most cases we won't probably know
anything about this backing volume but we need to in order to construct
a correct volume XML for libvirt. This patch was inspired by
virt-manager which (though in much more complexity) creates a transient
storage pool for user-specified backing store files to let libvirt
gather the necessary information for us to be populated back to libvirt
in an XML for a new volume to be created.

Signed-off-by: Erik Skultety <eskultet@redhat.com>