install: Don't create a symlink to the base image in the storage pool
This is a remnant from an earlier revision of the code where originally
the symlink was supposed to always point to the latest vendor cloud
image. However, as it was pointed out in the review at that time, that
would never work, because multiple machines might be using the same
backing image concurrently and so we'd affect VM instances which we
would not really intend to do when switching to usage of a newer vendor
cloud base image. Therefore, drop the related code and simply always
point to the actual base image located in lcitool's cache.

Signed-off-by: Erik Skultety <eskultet@redhat.com>