libvirt_wrapper: Add a simple helper to create a transient 'dir' pool
The idea behind this trivial helper was inspired by virt-manager and in
essence should allow us to fetch data about a backing store volume
using libvirt which we'd otherwise have to do ourselves. The pool is
supposed to be destroyed by the caller when the backing store volume is
processed.

Signed-off-by: Erik Skultety <eskultet@redhat.com>