config: Permit no user config on VM installs
Commit dc74805e seemingly allowed usage of lcitool's install feature
without providing any user config and instead rely on default values.
That's nice and it kinda works, but only when an actual config.yml file
exists. If it doesn't, we just return from the config loading method
returning None which is wrong. This patch fixes it properly, so users
truly don't need to provide a config.yml if our defaults are enough.
A corresponding test case is provided this time.

Fixes: dc74805efebd43a2120a26499822641c70e96fc0

Signed-off-by: Erik Skultety <eskultet@redhat.com>