inventory: Skip unknown target OS groups instead of failing immediately
If a user doesn't update their host inventory whenever we drop support
for a distro in lcitool OR they provide their own target OS definitions
via '--datadir', they're presented with the following (cryptic) error:

    Inventory error: Host 'foo' not found in any target OS group

What the error actually says is that we found a definition of a target
either in the Ansible inventory in form of a group or one which in
previous invocations came from user's '--datadir' that we don't know.
Either way, it has nothing to do with the host 'foo'.
Fix this by checking early for unsupported target OS groups found in
any inventory source. Given that the user is free to provide '--datadir'
on the commandline as they want/need, we must not make this kind of
error a fatal one, because it's perfectly valid to install a local VM
with a local target OS override residing in '--datadir' while not
making use of it all the time and instead default to regular lcitool
CLI usage, where the target OS definitions come from our own repo data.
Instead, make this a silent 'info' type of message, since we don't have
enough context on what the user intended to do, so not even a warning
is justified here.

Signed-off-by: Erik Skultety <eskultet@redhat.com>