containers: Only bind /etc/{passwd,group} when run as non-root
Utilities like passwd, useradd, etc. can't be used when we bind mount
the files inside the container. This is especially problematic on some
distros, e.g. Debian 11 slim, which expect to have a dedicated user
(_apt) present to execute a package upgrade; or a specific group
present to configure a package in a post scriptlet. If the user
requests running as root inside the container we should not restrict
him at all by bind mounting these files.
As for regular users inside the container nothing changes, IOW if root
is required inside the container lcitool needs to be instructed
accordingly, sudo wouldn't work because of the reasons described above
even if it were somehow available in the container.

There's another change in parallel with fixing the above in order to
make the tests pass again. As the problem described above was fixed by
making some of the docker/podman CLI options to only be passed when
needed, suddenly the expected test data no longer corresponded with the
actual output, because the options were in a different order which is
impossible to simulate in tests easily. So, what we need to make the
tests pass again is to sort the CLI options, but then we end up with a
weird result hard to debug, because options that take an argument would
out of the logical order. Therefore, all the container options are
passed around as a list of tuples for this particular reason - to fix
ordering - and is converted to a plain list of options right before
making the call to subprocess (which would not be able to handle tuples
inside the list).

Signed-off-by: Erik Skultety <eskultet@redhat.com>