lcitool: ansible: update: Add a 'changed_when' when checking Python
Because we have to use the 'command' module for this, idempotency isn't
available, so this tasks always produces a 'changed' status for the
task even though we're just capturing CLI output from shell.
'changed_when: false' means that this task will never yield a 'changed'
state. This is just a pre-caution measure rather than a bug fix.

Signed-off-by: Erik Skultety <eskultet@redhat.com>