targets: track the native architecture explicitly
Currently when generating package lists we assume the native
architecture is the same as the current host architecture. This
change decouples that assumption allowing native architecture to
be determined explicitly.

This fixes the 'manifest' command so that its output always
targets 'x86_64' as the native architecture, regardless of what
host it is run on. This is neccessary since GitLab shared runners
are 'x86_64'.

In future we we extend the 'manifest' command to generate configs
suitable for custom runners, then we would need to expand the
manifest.yml format to allow specifying a list of runners and
their native arches.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>