ansible: Bump the minimum Ansible version to 2.10+
It's getting harder to get around with only the builtin modules and
we'll need to use some of the curated community modules. The split of
Ansible into 'ansible-core' and 'ansible community (aka ansible)'
happened in v2.10. The way this patch enforces the version used in by
delegating a version check task to localhost, failing if the control
node doesn't meet the minimum version. 2.10 is pretty common these
days, so essentially only users on Debian-10 would be impacted by
this, but then again, installing from PyPI is a simple workaround in
that case.

Signed-off-by: Erik Skultety <eskultet@redhat.com>