ansible: update: system: Use 'dnf' module to handle Rawhide
The rest of the RPM distros are making use of the 'package' module
already, but because we needed to special case Rawhide in some aspects,
we've historically been using the simple 'command' module for it. With
the special casing now mostly gone, we can convert the plain 'command'
usage to the 'dnf' module.

Signed-off-by: Erik Skultety <eskultet@redhat.com>