ansible: system_update: Use pkgng module to install/update pgks on BSD
As with the other distros using native Ansible packager modules we're
aiming for idempotency here too. However, with pkgng isn't part of the
builtin collection of modules maintained by the core Ansible team.
Instead, it's been moved to the currated community collection which
means users of lcitool will need to make sure they have the collection
installed - which is quite simple, because on most distros it is
achieved by installing the 'ansible' package which will also implicitly
pull in the core 'ansible-core' package. Users relying on the
requirements.txt file will not notice any change.

There was a severe bug in the Ansible 2.9 version of the module (i.e.
before the [Ansible/ansible-core, ansible community] split) where
instead of updating all packages using the '*' syntax the module would
install every package in the repo, hence we bumped the Ansible version
to 2.10+ in a past patch.

Signed-off-by: Erik Skultety <eskultet@redhat.com>