ansible: users: Provide a static salt for user passwords
If we don't define any salt, the 'user' module naturally can't be
idempotent as the password hash will always be different, reporting a
'changed' status while no functional change was performed. Since
knowing the salt should not decrease strength of the resulting hash nor
help finding the plain text password (not that we'd really care in
lcitool's use cases). Therefore, specify a salt string so that the
password is changed exactly once.

Signed-off-by: Erik Skultety <eskultet@redhat.com>