ansible: base: Make sure python3-dnf is installed on Rawhide
Rawhide is expect to default to DNF5 soon. DNF5 decoupled the 'dnf'
binary from the library so that the binary can live without the
library/API. IOW, up until the point of the switch the 'dnf' package provided a symlink to
the 'dnf-3' binary provided by the 'python3-dnf' package which means
that both were always installed together with the latter marked as
protected by DNF.
However, in order to use either the 'dnf' builtin Ansible module or the
generic 'package' one which translates to 'dnf', we need to be able to
communicate with DNF on the managed node. The problem is that the 'dnf'
Ansible module relies on the 'python3-dnf' package which DNF5 doesn't
need nor use, hence we need to make sure it's installed on Rawhide even
though it only provides the "legacy" functionality compared to DNF5.

Note that even when Rawhide (or regular Fedora for that matter)
switches to DNF5 we still won't be able to use the 'dnf5' Ansible
module at least for some time, because that module:
    - is still in development and bound to change
    - was introduced in ansible-core v2.15 which isn't widely available
      at the time of this patch (which may change by the time DNF5 is
      mature)

Signed-off-by: Erik Skultety <eskultet@redhat.com>