ansible: system_update: Add a comment on the Rawhide GPG keys update
It's not obvious at first glance why we need to update GPG keys on
Rawhide or why we perform the update there in multiple steps. One has
to dig into the commit history to find the reasoning in there.
Extract the gist from the commit messages and put commentaries to the
relevant piece of code.

Signed-off-by: Erik Skultety <eskultet@redhat.com>