ansible: system_update: Enable CRB/PowerTools repos idempotently
The Ansible 'dnf' module doesn't have support for the 'config-manager'
subcommand despite the community asking for it, so we have to look for
a different way of enabling an already installed but potentially
disabled repo. We can use one of the curated community modules part of
the "new" ansible package - ini_file. In order to make it clear that
this is not a builtin module from ansible-core, use FQDN instead of a
simple module name.

Signed-off-by: Erik Skultety <eskultet@redhat.com>