ansible: system_update: [BSD] Clean up packages idempotently
Since we're already making use of the 'pkgng' module, use it to
autoremove leaf packages on BSD too. Same as with the rest of distros,
cleaning the package cache isn't covered by any module and also doesn't
provide any (useful) output, so that particular task needs to be
handled with a 'changed_when' clause to ensure idempotency.

Signed-off-by: Erik Skultety <eskultet@redhat.com>