ansible: system_setup: [Fedora/CentOS] Clean up packages idempotently
Clean the packages post-update in an idempotent way.
Firstly, there's no way how to make package cache clean up idempotent
because no module actually has that feature. However, we don't really
expect any change from this module other than wiping some internal
structures, so pairing the 'command' module with a 'changed_when:
false' is trivially going to achieve idempotency.

Secondly, we sadly have to special-case Rawhide in terms of
autoremoving leaf packages by using a plain 'command' instead of the
'dnf' module for that. This is because Rawhide has recently switched to
DNF5 which is different to DNF which for purposes of the Ansible module
defaults to the 'dnf-3' binary provided by the python3-dnf package. In
other words if we used the 'command' module to install 'python3-dnf' on
Rawhide just to be able to use the 'dnf' Ansible module in subsequent
tasks, 'dnf-3' will think that the 'python3-dnf' package is leaf and
would try to autoremove it, but fail because it's protected. This is
all because different DNF binaries are being used in different stages
of the playbook on Rawhide at the moment. conversely, if we use
'command' on Rawhide, it'll resort to the same DNF5 binary which was
used to install 'python3-dnf' in the beginning and hence won't try to
autoremove it.

Signed-off-by: Erik Skultety <eskultet@redhat.com>