util: Add a helper to query inventory from datadir
We'll enable support for inventory overrides in future patches, IOW let
users to specify a different location for their inventory other than
the default under $HOME/.config/lcitool. While having a --inventory
option is certainly possible, given that we don't expose other Ansible
options directly on the command line and instead opted for the usage of
a data dir, let's try loading an inventory override from that location
as well.

Signed-off-by: Erik Skultety <eskultet@redhat.com>