inventory: Add an argument to specify inventory override
Officially add support for inventory overrides to the object
constructor. While this might not be a super useful feature on its own
at the moment, the true value here is that we'll be able to get rid of
some ugly pytest monkeypatched fixtures, including one for the
inventory.
The choice of the argument 'inventory_path' may not be the best since
Ansible actually supports a list of inventory sources that it can load
data from, but that would require more serious refactor which isn't
really needed at this point in time since we haven't converted lcitool
to a library yet, so we can change it any time.

Signed-off-by: Erik Skultety <eskultet@redhat.com>