tests: Replace the monkeypatched inventory fixture with a proper one
The reason for the monkeypatched inventory fixture was that we didn't
have a way of passing an override inventory source to the object
constructor and also because we called into libvirt to query the list
of hosts to extend the inventory. Now we do have a way to pass such an
override which means libvirt querying will be skipped automatically and
so we don't need the monkeypatched version of the fixture, hence we can
replace it with a nicer one.

Signed-off-by: Erik Skultety <eskultet@redhat.com>