config: Support passing in a custom config file to load
We're limiting users to create a config file in their home directory
under $HOME/.config/lcitool or under $XDG_HOME_DIR. What if all they
need is a quick smoke test with varying configuration files? In that
case creating directories and setting the env variable is cumbersome
and there isn't really a reason why we shouldn't have a --config CLI
option. This patch lays the foundation for that on the interface level.
As a nice side-effect this will also allow us to drop a very ugly
pytest config monkeypatch fixture we have.

Signed-off-by: Erik Skultety <eskultet@redhat.com>