mappings: Add awk
Autoconf requires awk to convert config.h.in into config.h; but
OpenSUSE Tumbleweed fails to list awk as a prerequisite of autoconf,
leading to failures like this in libnbd:
https://gitlab.com/libvirt/libvirt-ci/-/issues/28
 configure: creating ./config.status
 config.status: creating podwrapper.pl
 ./config.status: line 1496: awk: command not found

As a first step, list awk in the mappings, so that projects can
request it, and regenerate the unit tests to match.  Here, we are
preferring the 'awk' that distros favor by default; we could add a
separate rule for gawk if GNU awk is specifically needed.

Signed-off-by: Eric Blake <eblake@redhat.com>