ansible: system_update: Drop updating packages in the Rawhide block
Now that we're updating the GPG keys package using the 'dnf' module
with package cache being refreshed during that task we don't need to
use a dedicated package update task for Rawhide and instead we can
re-use the already existing RPM distro package update task. The trick
is to move that task after the one that updates the GPG keys on Rawhide
and hence optimize the code a bit.

Signed-off-by: Erik Skultety <eskultet@redhat.com>