ansible: system_update: Drop explicit usage of Rawhide nodebug kernel
Rawhide has defaulted to the nodebug kernel for a while now, leaving
the debug kernel only as optional. This means we can simplify the
update of packages in Rawhide and we no longer need to perform it in
two steps. This patch also removes the nodebug kernel repository since
both the debug and nodebug kernels are available via the mainstream
updates repo which is updated almost on a daily basis as well.

Signed-off-by: Erik Skultety <eskultet@redhat.com>