application: Rename _get_client to _container_handle
The original name was a bit misleading in that we don't know what
client is. The new name clearly demonstrates that whatever object we
get back can be used to manipulate the engine below.
Along with those changes take the opportunity to replace all 'client'
variable references with 'engine' for the very same reason.

Signed-off-by: Erik Skultety <eskultet@redhat.com>