application: Extract the common container bits into separate helper
We set a bunch of options for the underlying container engine through
CLI options. These are almost completely identical for both a generic
command execution as well as with executing an interactive shell.
Future patches will add a dedicated _action_X entry point for the
'shell' use case which means we'll have to set the options in a shared
helper private method this patch introduces.

Signed-off-by: Erik Skultety <eskultet@redhat.com>