application: Catch ContainerExecError on the application level
So far all container operations are interactive which means that when a
failure occurs we'll see in in the terminal, because parent's stdout
and stderr will be pipe-connected to the container backend processes.
In case of an error an extra ApplicationError exception would be raised
with the current code. Instead, we should skip reporting any redundant
exception on the application level and just exit with the child
process's return code. We'll achieve that by catching
ContainerExecError on the global application level.

Signed-off-by: Erik Skultety <eskultet@redhat.com>