lcitool: add missing function call in error message
When constructing a list of hosts managed by lcitool, we try to obtain
lcitool metadata from each libvirt domain. If we fail to obtain metadata
LibvirtWrapperError with the domain name is raised. However, there is a
typo, where the function is included in the error message instead of the
result of a function call.

Signed-off-by: Daniel Zatovic <daniel.zatovic@gmail.com>