lcitool: fix obtaining transient domain metadata
During the update process, lcitool builds ansible inventory by iterating
through all libvirt domains, looking for the ones with lcitool metadata.
It obtains the metadata of each host using the VIR_DOMAIN_AFFECT_CONFIG
flag. However, if there is at least one transient domain running on the
system, the loop stops and lcitool exits with an error (the transient
domain can be totally unrelated to lcitool and not managed by it). We
fix this by checking if the domain is not persistent (i.e. transient).
If so, we obtain the live config (VIR_DOMAIN_AFFECT_LIVE), otherwise we
fall-back to the previous behaviour (i.e. use VIR_DOMAIN_AFFECT_CONFIG).

Signed-off-by: Daniel Zatovic <daniel.zatovic@gmail.com>