mappings: Add gettext-native
Like glib2-native, this is going to be useful for scenarios in
which we're cross-building but still need to run the native
version of one of the gettext tools as part of the build.

Things already work fine for most cross-build targets, but for
MinGW we have an override that results in only the foreign
package being installed.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>