manifest: merge 'prebuilt-env' and 'local-env' jobs together
In the previous major overhaul of CI jobs to better support merge
request pipelines, we ended up duplicating all build jobs with
"prebuilt-env" and "local-env" suffixes. The former used a cached
container image with all packages pre-installed, while the latter
used the target OS based image and freshly installed the packages.

It was pointed out that it is possible to set job variables from
within a 'rules' cause and such variables will be accessible even
in the 'image:' tag. This makes it possible to have a single job
that references the pre-built container image, and write rules
that override this to point to the target base image.

Thus we can now eliminate the annoying duplication of jobs to
increase our sanity.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>