mappings: Switch the bash-completion to bash-completion-devel
While it may seem that bash-completion was split into two only in
Fedora rawhide, that's not the entire truth. Firstly, this change
was done in OpenSUSE distros ages years ago [1][2], but because
of the default we're still installing just bash-completion. This
results in Libvirt not detecting the corresponding pkg file
(rightfully though) [3].

But instead of listing overrides for Fedora rawhide and OpenSUSE
Leap and Tumbleweed let's switch the default and explicitly list
'older' distros (like CentOS, AlmaLinux, older Fedora) as these
overrides will be removed progressively as we drop those distros.

1: https://build.opensuse.org/package/view_file/openSUSE:Leap:15.0/bash-completion/bash-completion.spec?expand=1
2: https://build.opensuse.org/package/view_file/openSUSE:Factory/bash-completion/bash-completion.spec?expand=1
3: https://gitlab.com/libvirt/libvirt/-/jobs/6258062679#L806

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>