List installed packages on Alpine only
After a container is done installing necessary packages, the list
of all installed packages is collected into packages.txt file
which is then printed out. And this works everywhere but Alpine.
Because there 'apk list' is run which prints ALL available
packages. We need to pass '--installed' to limit listing only to
the installed ones [1].

1: https://man.archlinux.org/man/apk-list.8.en

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>