lcitool: fix running jobs when there are CI changes present
If using a 'when: never' clause, evaluation of any following rules is
skipped. We attempted to use 'when: never' to skip running merge
request pipelines with pre-built containers, when there are CI changes
present. Unfortunately evaluation will not carry on and catch the rule
that runs CI with one-off built containers.

There is no way to invert the meaning of a 'changes:' clause, so the
only way to achieve what we want is to re-arrange the rules, so that
we can avoid use of 'when: never'.

The current grouping of rules is based on use of pre-built containers
vs one-off built containers. This is proving hard to analyse as the
rules for 'push' events are far apart from each other.

This changes the grouping to be based on the event type, so everything
that is conceptually related to the same event as in one place. With
this grouping, we simply put rules with filename matches first, and
get the desired ordering semantics.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>