mappings: Make some more packages available on CentOS Stream 9 / AlmaLinux 9
A bunch of packages are currently marked as non-available in CentOS
Stream 9 / AlmaLinux 9, though they are available in the distribution
(or in EPEL 9). Looks like these were originally disabled since they
were not supported in CentOS 7, then only enabled for CentOS Stream 8
and AlmaLinux 8, but then forgotten to be added when the major version 9
came around.

Since we already dropped support for CentOS 7 in lcitool a while ago,
we can simply fix the problem by removing the special entries for
generic CentOS / AlmaLinux and for version 8 of these distributions,
so that all of them now use the default "rpm:" mapping.

Signed-off-by: Thomas Huth <thuth@redhat.com>