mappings: Fix genisoimage related entries for CS9 / AlmaLinux 9
genisoimage is available on the EL 9 distros, too. Since we do not
have to take care of CentOS 7 anymore, we can now simply add a
generic "rpm:" entry for this which will cover all, Fedora, CentOS
and AlmaLinux. (OpenSUSE has a separate entry anyway which will
prevent that this distribution is affected by the new "rpm:" line).

Signed-off-by: Thomas Huth <thuth@redhat.com>