containers: introduce 'cirrus-vars' command
The gitlab CI config file has to pass various variables into the cirrus
CI config file to control the build process. This is done by a giant sed
command in the cirrus build template. This is inflexible since the set
of substitutions is hardcoded.

This new 'cirrus-vars' command can replace any variable @NAME@ with the
corresponding value from the $NAME environment variable.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>