gitlab: use 'cirrus-vars' tool for generating build file
The giant 'sed' command that is used to generate a Cirrus CI build YAML
file has a hardcoded list of environment variable substitutions it can
perform. For most projects this list is overkill and for some projects
it is insufficient.

Now that the 'cirrus-run' container has the 'cirrus-vars' command, we
can invoke that to dynamically perform the substitutions. Since the
'vars' files don't export the shell variables they declare, we must
add 'set -o allexport' to force all vars to be implicitly exported.

Note that when updating projects across this commit, their 'build.yml'
file will need changing for the special handling of '$PATH' that used
to be done by sed, and must now be done in the 'build.yml' directly:

  -  PATH: "@PATH@"
  +  PATH: "@PATH_EXTRA@:$PATH"

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>