lcitool: default unreleased distros to 'allow-failure: true'
In recent times we have increasingly been marking the unreleased
distros (rawhide, tumbleweed, sid, etc) as 'allow-failure: true'
in many projects. This reflects the fact that they all suffer
from periods in time when packages are not reliably installable,
or those that are installed are broken in some way.

Rather than continue to mark individual projects, this adds an
'stable: true|false' field in the target yaml files, with all
the unreleased distros being marked 'stable: false'.

Any target with 'stable: false' will get 'allow-failure: true'
set by default. The project's manifest.yml file can still
override this setting it back to 'allow-failure: false', if a
project really wants to treat an unreleased target as gating,
though this is not recommended.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>