mappings: Reflect libclang-dev split on Ubuntu2404
In Ubuntu 24.04 (well, 23.04 even) the libclang-dev package was
renamed to libclang-rt-dev (see [1] and search for
1:14.0.6-10~exp4 release). Since this is basically a new package,
add it as such to our mappings file. And of course, some distros
(like Alpine, OpenSUSE) did not make that change, or some older
ones (like Ubuntu 22.04 or Debian 11) are too old for this
change.

NB, some distros include this under 'Recommended packages', but
usually we pass arguments to apt-get/dnf/... that make it not
install those.

1: https://changelogs.ubuntu.com/changelogs/pool/main/l/llvm-toolchain-18/llvm-toolchain-18_18.1.3-1/changelog
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>