tests: drop support for almalinux-8
almalinux-8 is going away since CentOS 8 EOL prevents use of the
advanced virtualization repository.

While we could change the YUM config to point to vault.centos.org,
rather than the mirror server, libvirt & QEMU both have a 2 year
cut off for support of the previous major version of a distro, so
it is not beneficial to invest time to keep almalinux 8 working.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>